requirements.txt
lmdb
pyclipper
shapely
addict
matplotlib
addict
numpy
tqdm>=4.64.1
opencv-python-headless>=3.4.18.65
rapidfuzz>=2.13.7
Pillow>=9.1.1
PyYAML>=6.0
xml-python>=0.4.3
packaging
Cython
lanms>=1.0.2
